+++ tests/CurrentProcessTest.php
@@ -13,6 +13,7 @@ use Innmind\Server\Control\Server\Process\Pid;
 use Innmind\Server\Status\Server\Memory\Bytes;
 use Innmind\TimeContinuum\Period;
 use Innmind\TimeWarp\Halt;
+use Innmind\Signals\Handler;
 use Innmind\Immutable\SideEffect;
 use Psr\Log\NullLogger;
 use Innmind\BlackBox\{
@@ -27,7 +28,10 @@ class CurrentProcessTest extends TestCase
 
     public function testId()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Pid::class, $process->id()->unwrap());
         $this->assertSame(
@@ -57,7 +61,10 @@ class CurrentProcessTest extends TestCase
 
     public function testSignals()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Signals::class, $process->signals());
         $this->assertSame($process->signals(), $process->signals());
@@ -65,7 +72,10 @@ class CurrentProcessTest extends TestCase
 
     public function testMemory()
     {
-        $process = CurrentProcess::of(Halt\Usleep::new());
+        $process = CurrentProcess::of(
+            Halt\Usleep::new(),
+            Handler::main(),
+        );
 
         $this->assertInstanceOf(Bytes::class, $process->memory());
         $this->assertGreaterThan(3_000_000, $process->memory()->toInt()); // ~3MB